refactor(v0.14.0): on-demand pre-bundle
